Create tests
The code you fill in the C code files should at the end of its execution print to standard output, on each line, the ID of the process running at each time (starting at time 0). For the processes defined in the tests/INPUT01 file the output that should be generated will be the one found in the corresponding tests/fcfs01, tests/sjf01, tests/srtf01 and tests/rr01 files.